Title: 나눗셈의 나머지, Time: 98ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240519/나눗셈의 나머지/remainder-of-division.py
@@ -2,7 +2,7 @@ a,b=tuple(map(int,input().split()))
 tmp=[0]*b
 while(a>1):
     tmp[int(a)%int(b)]+=1
-    a=a/b
+    a=int(a/b)
 ans=0
 for i in tmp:
     ans+=(i**2)
